refactor: refactor helper functions
