+++ awsmt/data_source_playback_configuration.go
@@ -167,81 +167,3 @@ func dataSourcePlaybackConfigurationRead(_ context.Context, d *schema.ResourceDa
 	}
 	return diags
 }
-
-func getSinglePlaybackConfiguration(c *mediatailor.MediaTailor, name string) (*mediatailor.PlaybackConfiguration, error) {
-	output, err := c.GetPlaybackConfiguration(&mediatailor.GetPlaybackConfigurationInput{Name: &name})
-	if err != nil {
-		return nil, err
-	}
-	return (*mediatailor.PlaybackConfiguration)(output), nil
-}
-
-func flattenPlaybackConfiguration(c *mediatailor.PlaybackConfiguration) map[string]interface{} {
-	if c != nil {
-		output := make(map[string]interface{})
-
-		output["ad_decision_server_url"] = c.AdDecisionServerUrl
-		if !(*c.AvailSuppression.Mode == "OFF" && c.AvailSuppression.Value == nil) {
-			output["avail_suppression"] = []interface{}{map[string]interface{}{
-				"mode":  c.AvailSuppression.Mode,
-				"value": c.AvailSuppression.Value,
-			}}
-		}
-		if c.Bumper.EndUrl != nil || c.Bumper.StartUrl != nil {
-			output["bumper"] = []interface{}{map[string]interface{}{
-				"end_url":   c.Bumper.EndUrl,
-				"start_url": c.Bumper.StartUrl,
-			}}
-		}
-		if c.CdnConfiguration.AdSegmentUrlPrefix != nil || c.CdnConfiguration.ContentSegmentUrlPrefix != nil {
-			output["cdn_configuration"] = []interface{}{map[string]interface{}{
-				"ad_segment_url_prefix":      c.CdnConfiguration.AdSegmentUrlPrefix,
-				"content_segment_url_prefix": c.CdnConfiguration.ContentSegmentUrlPrefix,
-			}}
-		}
-		if c.ConfigurationAliases != nil {
-			output["configuration_aliases"] = c.ConfigurationAliases
-		}
-		output["dash_configuration"] = []interface{}{map[string]interface{}{
-			"manifest_endpoint_prefix": c.DashConfiguration.ManifestEndpointPrefix,
-			"mpd_location":             c.DashConfiguration.MpdLocation,
-			"origin_manifest_type":     c.DashConfiguration.OriginManifestType,
-		}}
-		output["hls_configuration"] = []interface{}{map[string]interface{}{
-			"manifest_endpoint_prefix": c.HlsConfiguration.ManifestEndpointPrefix,
-		}}
-		if !(c.LivePreRollConfiguration.MaxDurationSeconds == nil && c.LivePreRollConfiguration.AdDecisionServerUrl == nil) {
-			output["live_pre_roll_configuration"] = []interface{}{map[string]interface{}{
-				"ad_decision_server_url": c.LivePreRollConfiguration.AdDecisionServerUrl,
-				"max_duration_seconds":   c.LivePreRollConfiguration.MaxDurationSeconds,
-			}}
-		}
-		if c.LogConfiguration != nil {
-			output["log_configuration"] = []interface{}{map[string]interface{}{
-				"percent_enabled": c.LogConfiguration.PercentEnabled,
-			}}
-		} else {
-			output["log_configuration"] = []interface{}{map[string]interface{}{
-				"percent_enabled": 0,
-			}}
-		}
-		if *c.ManifestProcessingRules.AdMarkerPassthrough.Enabled {
-			output["manifest_processing_rules"] = []interface{}{map[string]interface{}{
-				"ad_marker_passthrough": []interface{}{map[string]interface{}{
-					"enabled": c.ManifestProcessingRules.AdMarkerPassthrough.Enabled,
-				}},
-			}}
-		}
-		output["name"] = c.Name
-		output["personalization_threshold_seconds"] = c.PersonalizationThresholdSeconds
-		output["playback_configuration_arn"] = c.PlaybackConfigurationArn
-		output["playback_endpoint_prefix"] = c.PlaybackEndpointPrefix
-		output["session_initialization_endpoint_prefix"] = c.SessionInitializationEndpointPrefix
-		output["slate_ad_url"] = c.SlateAdUrl
-		output["tags"] = c.Tags
-		output["transcode_profile_name"] = c.TranscodeProfileName
-		output["video_content_source_url"] = c.VideoContentSourceUrl
-		return output
-	}
-	return map[string]interface{}{}
-}
